# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Skorbord Card Scoring API
  description: |
    REST API for the Skorbord card scoring application. This API provides endpoints for managing players, games, scores, and real-time updates via WebSocket connections.
    
    ## Features
    - Player management (CRUD operations)
    - Game creation and management
    - Real-time scoring with WebSocket support
    - Game statistics and rivalry tracking
    - Multiple game type support
    
    ## Real-time Updates
    The API supports real-time updates via Socket.IO. Connect to the WebSocket endpoint to receive:
    - Live score updates
    - Game status changes
    - Player updates
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse:
    - 100 requests per 15 minutes per IP address
    - Stricter limits may apply to certain endpoints
    
  version: 1.0.0
  contact:
    name: Skorbord Support
    url: https://skorbord.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://skorbord.app/api
    description: Production server
  - url: http://localhost:2424/api
    description: Development server
paths:
...
  /games:
    post:
      summary: Create a new game
      description: Create a new game session with specified players and settings
      operationId: createGame
      tags:
        - Games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - game_type_id
                - player_ids
              properties:
                game_type_id:
                  type: string
                  description: ID of the game type to play
                player_ids:
                  type: array
                  items:
                    type: string
                  minItems: 2
                  description: Array of player IDs participating in the game
                settings:
                  type: object
                  description: Game-specific settings
                  additionalProperties: true
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameWithPlayers'
        '400':
          $ref: '#/components/responses/ValidationError'
  /games/{id}:
    get:
      summary: Get game details
      description: Retrieve complete game information including players, scores, and rounds
      operationId: getGame
      tags:
        - Games
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    GameId:
      name: id
      in: path
      required: true
      description: Unique identifier for a game
      schema:
        type: string
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
                    value:
                      oneOf:
                        - type: string
                        - type: number
                        - type: boolean
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    GameType:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
        name:
          type: string
          description: Display name of the game
        description:
          type: string
          description: Detailed description
        min_players:
          type: integer
          description: Minimum number of players
        max_players:
          type: integer
          description: Maximum number of players
        supports_teams:
          type: boolean
          description: Whether the game supports team play
        default_settings:
          type: object
          description: Default game settings
          additionalProperties: true
      required:
        - id
        - name
        - description
        - min_players
        - max_players
        - supports_teams
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
        short_id:
          type: string
          description: Short alphanumeric code for easy sharing
        game_type_id:
          type: string
          description: ID of the game type
        status:
          type: string
          enum: [active, paused, completed, cancelled]
        settings:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - short_id
        - game_type_id
        - status
        - settings
        - created_at
        - updated_at
    GameWithPlayers:
      allOf:
        - $ref: '#/components/schemas/Game'
        - type: object
          properties:
            players:
              type: array
              items:
                type: object
                properties:
                  player_id:
                    type: string
                  name:
                    type: string
                  avatar_color:
                    type: string
                  position:
                    type: integer
    GameDetails:
      allOf:
        - $ref: '#/components/schemas/GameWithPlayers'
        - type: object
          properties:
            game_type:
              $ref: '#/components/schemas/GameType'
            scores:
              type: array
              items:
                type: object
                properties:
                  player_id:
                    type: string
                  total_score:
                    type: integer
                  round_scores:
                    type: array
                    items:
                      type: integer
            rounds:
              type: array
              items:
                $ref: '#/components/schemas/Round'
    Round:
      type: object
      properties:
        id:
          type: string
        game_id:
          type: string
        round_number:
          type: integer
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        scores:
          type: array
          items:
            type: object
            properties:
              player_id:
                type: string
              points:
                type: integer
      required:
        - id
        - game_id
        - round_number
        - completed
        - created_at
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code (optional)
        details:
          type: object
          description: Additional error details (optional)
          additionalProperties: true
      required:
        - error
...